projects/ml_engineer_project/requirements.txt
# Core ML Libraries
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
imbalanced-learn==0.11.0

# API & Web
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# MLOps
mlflow==2.7.1
prometheus-client==0.17.1

# Data Drift & Monitoring
scipy==1.11.2
evidently==0.4.5

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Database
psycopg2-binary==2.9.7
redis==5.0.0

# Cloud
boto3==1.28.25

# Testing
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1

# Code Quality
black==23.7.0
pylint==2.17.5
